test(export): move the TensorRT-LLM-only export test into tests/.../export/trtllm
Mirror the source split on the test side: tests that exercise code now under
modelopt/torch/export/trtllm live in a matching trtllm/ test package.

Only test_get_scaling_factor_from_weight qualified. It covers
get_scaling_factor_from_weight, which moved to trtllm/quant_utils.py because
trtllm/postprocess.py is its only caller. Every other test in
tests/gpu/torch/export/test_export.py exercises the backend-agnostic
quant_utils helpers (process_layer_quant_config is shared with
unified_export_megatron, and the final test drives export_hf_checkpoint), so
they stay put.

tests/gpu_trtllm/ is deliberately left alone: that tier means "needs
TensorRT-LLM installed", not "tests TensorRT-LLM export".
test_export_compressed_nvfp4.py lives there because it needs
torch.ops.trtllm.fp4_quantize to produce a cutlass-swizzled scale, but it
tests unified_export_hf, which is the HF path.

The new module is named test_trtllm_quant_utils.py rather than
test_quant_utils.py: pytest runs these directories without __init__.py and
derives the module name from the bare filename, so the shorter name collides
with tests/gpu/torch/export/test_quant_utils.py and fails collection with
"import file mismatch". The docstring records the convention.

Testing: pytest collects tests/gpu/torch/export cleanly (172 items, no
collection error) and the moved test passes (4 parametrized cases).
tests/unit/torch/export plus the new directory: 193 passed. pre-commit passes
on both changed files.
